requirements.txt
tensorflow-gpu==2.3
tensorboard==2.3
tensorboard_plugin_profile==2.3
matplotlib==3.3.2
scikit-learn==0.23.2
scipy==1.4.1
pybind11==2.4.3
pytest==6.1.1
opencv-contrib-python==4.4.0.46
gdown==3.12.2
